requirements.txt
pandas
pypandoc
pandoc-xnos
git+https://github.com/olivierverdier/pydflatex#egg=pydflatex
blessings
doi2bib
ruamel.yaml
yaml
